Refactor to remove clone
