+++ ank/src/cli.rs
@@ -12,28 +12,9 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-use std::{error::Error, io};
-
-#[cfg(test)]
-use mockall::automock;
+use std::error::Error;
 
 use clap::{command, Parser, Subcommand};
-
-#[cfg(not(test))]
-pub fn open_manifest(
-    file_path: &str,
-) -> io::Result<(String, Box<dyn io::Read + Send + Sync + 'static>)> {
-    use std::fs::File;
-    match File::open(file_path) {
-        Ok(open_file) => Ok((file_path.to_owned(), Box::new(open_file))),
-        Err(err) => Err(err),
-    }
-}
-#[cfg(test)]
-use self::tests::open_manifest_mock as open_manifest;
-
-pub type InputSourcePair = (String, Box<dyn io::Read + Send + Sync + 'static>);
-pub type InputSources = Result<Vec<InputSourcePair>, String>;
 
 use common::DEFAULT_SERVER_ADDRESS;
 
@@ -226,145 +207,4 @@ where
 
 pub fn parse() -> AnkCli {
     AnkCli::parse()
-}
-
-#[cfg_attr(test, automock)]
-pub trait HandleApplyArgs {
-    fn get_input_sources(&self) -> InputSources;
-}
-
-// #[cfg_attr(test, automock)]
-impl HandleApplyArgs for ApplyArgs {
-    fn get_input_sources(&self) -> InputSources {
-        if let Some(first_arg) = self.manifest_files.first() {
-            match first_arg.as_str() {
-                // [impl->swdd~cli-apply-accepts-ankaios-manifest-content-from-stdin~1]
-                "-" => Ok(vec![("stdin".to_owned(), Box::new(io::stdin()))]),
-                // [impl->swdd~cli-apply-accepts-list-of-ankaios-manifests~1]
-                _ => {
-                    let mut res: InputSources = Ok(vec![]);
-                    for file_path in self.manifest_files.iter() {
-                        match open_manifest(file_path) {
-                            Ok(open_file) => res.as_mut().unwrap().push(open_file),
-                            Err(err) => {
-                                res = Err(match err.kind() {
-                                    io::ErrorKind::NotFound => {
-                                        format!("File '{}' not found!", file_path)
-                                    }
-                                    _ => err.to_string(),
-                                });
-                                break;
-                            }
-                        }
-                    }
-                    res
-                }
-            }
-        } else {
-            Ok(vec![])
-        }
-    }
-}
-
-//////////////////////////////////////////////////////////////////////////////
-//                 ########  #######    #########  #########                //
-//                    ##     ##        ##             ##                    //
-//                    ##     #####     #########      ##                    //
-//                    ##     ##                ##     ##                    //
-//                    ##     #######   #########      ##                    //
-//////////////////////////////////////////////////////////////////////////////
-#[cfg(test)]
-mod tests {
-    use std::io;
-
-    use crate::cli::{ApplyArgs, HandleApplyArgs, InputSourcePair};
-
-    mockall::lazy_static! {
-        pub static ref FAKE_OPEN_MANIFEST_MOCK_RESULT_LIST: std::sync::Mutex<std::collections::VecDeque<io::Result<InputSourcePair>>>  =
-        std::sync::Mutex::new(std::collections::VecDeque::new());
-    }
-
-    pub fn open_manifest_mock(
-        _file_path: &str,
-    ) -> io::Result<(String, Box<dyn io::Read + Send + Sync + 'static>)> {
-        FAKE_OPEN_MANIFEST_MOCK_RESULT_LIST
-            .lock()
-            .unwrap()
-            .pop_front()
-            .unwrap()
-    }
-
-    #[tokio::test]
-    async fn utest_apply_args_get_input_sources_manifest_files_error() {
-        let _guard = crate::test_helper::MOCKALL_CONTEXT_SYNC
-            .get_lock_async()
-            .await;
-
-        let _dummy_content = io::Cursor::new(b"manifest content");
-        FAKE_OPEN_MANIFEST_MOCK_RESULT_LIST
-            .lock()
-            .unwrap()
-            .push_back(Err(io::Error::other(
-                "Some error occurred during open the manifest file!",
-            )));
-
-        let args = ApplyArgs {
-            manifest_files: vec!["manifest1.yml".to_owned()],
-            agent_name: None,
-            delete_mode: false,
-        };
-
-        assert!(args.get_input_sources().is_err(), "Expected an error");
-    }
-
-    // [utest->swdd~cli-apply-accepts-ankaios-manifest-content-from-stdin~1]
-    #[test]
-    fn utest_apply_args_get_input_sources_valid_manifest_stdin() {
-        let args = ApplyArgs {
-            manifest_files: vec!["-".to_owned()],
-            agent_name: None,
-            delete_mode: false,
-        };
-        let expected = vec!["stdin".to_owned()];
-        let actual = args.get_input_sources().unwrap();
-
-        let get_file_name = |item: &InputSourcePair| -> String { item.0.to_owned() };
-        assert_eq!(
-            expected,
-            actual.iter().map(get_file_name).collect::<Vec<String>>()
-        )
-    }
-
-    // [utest->swdd~cli-apply-accepts-list-of-ankaios-manifests~1]
-    #[tokio::test]
-    async fn utest_apply_args_get_input_sources_manifest_files_ok() {
-        let _guard = crate::test_helper::MOCKALL_CONTEXT_SYNC
-            .get_lock_async()
-            .await;
-
-        let _dummy_content = io::Cursor::new(b"manifest content");
-        for i in 1..3 {
-            FAKE_OPEN_MANIFEST_MOCK_RESULT_LIST
-                .lock()
-                .unwrap()
-                .push_back(Ok((
-                    format!("manifest{i}.yml"),
-                    Box::new(_dummy_content.clone()),
-                )));
-        }
-
-        let args = ApplyArgs {
-            manifest_files: vec!["manifest1.yml".to_owned(), "manifest2.yml".to_owned()],
-            agent_name: None,
-            delete_mode: false,
-        };
-        let expected = vec!["manifest1.yml".to_owned(), "manifest2.yml".to_owned()];
-        let actual = args.get_input_sources().unwrap();
-
-        let get_file_name = |item: &InputSourcePair| -> String { item.0.to_owned() };
-        assert_eq!(
-            expected,
-            actual.iter().map(get_file_name).collect::<Vec<String>>()
-        )
-    }
 }
